feat(driver-adapters): extracted common types to "driver_adapters::types"
